Refactor user interaction code to src/index.js
The user interaction handling code has been moved from the main index.js file into src/index.js. The code to ask the user for the function they wish to run has been wrapped into a `askQuestion` function. In addition, the handling of errors has been improved to ensure that the correct information is displayed and the user is asked the question again in the event of an error.